kmeans_predict

+++ shrecsys/preprocessing/preKmeans.py
@@ -79,7 +79,7 @@ def build_users_embedding_np(videos_embedding, videos_index, view_seqs, with_use
         indices = np.where(tfidf > 0)
         rating = np.take(tfidf, indices)[0]
         videos_embed = np.take(videos_embedding_array, indices, axis=0)[0]
-        user_embedding = np.matmul(rating, videos_embed)
+        user_embedding = np.matmul(rating, np.float64(videos_embed))
         users_embedding.append(user_embedding)
         if i % 1 == 0:
             logging.info("build users embedding, users index:{}".format(i))
